refactor(view): move client logic to separate component and clean up unused files
- Extract client-side logic from view page into dedicated view-client component
- Remove unused test API route and supabase import from raffle-prizes route
