+++ src/app/api/raffle/raffle-prizes/route.js
@@ -1,4 +1,4 @@
-import { createClient } from "@/utils/supabase/server"
+
 import { NextResponse } from "next/server"
 
 export async function POST(req) {
